- updated styling and converted js file to ts for react componnets
