+++ docs/src/assets/styles/variables.scss
@@ -6,5 +6,6 @@ $primarycolor: #1d232f;
 $secondarycolor: #2770ef;
 $lightgrey: #eaedf2;
 $darkgrey: #777e8b;
+$borderColor: #c0c6cf;
 $hovercolor: rgba(113, 161, 244, 0.12);
 $disabled-color: #a5acb9;
