+++ docs/src/components/BackButton/index.scss
@@ -4,7 +4,7 @@
     display: flex;
     padding-left: 24px;
     align-items: center;
-    
+
     button {
         width: 40px;
         height: 40px;
@@ -19,5 +19,6 @@
     }
     p {
         color: $darkgrey;
+        font-size: 14px;
     }
 }
