+++ docs/src/components/Docmap/index.scss
@@ -7,6 +7,7 @@
     position: fixed;
     right: 0;
     height: 100%;
+    font-size: 14px;
 
     ul {
         list-style-type: none;
