+++ docs/src/components/Document/index.scss
@@ -1,6 +1,6 @@
 .documentWrapper {
     width: calc(100% - 300px);
-    padding: 24px;
+    padding: 60px;
 }
 
 @media screen and (min-width: 767px) and (max-width: 1024px) {
