+++ docs/src/components/LeftSidebar/index.scss
@@ -13,13 +13,16 @@ aside {
     margin-top: 50px;
 
     .heading {
-        padding-left: 24px;
+        padding: 30px 0px 30px 24px;
+        margin: 0;
+        font-size: 18px;
     }
 
     .subHeading {
         margin: $gutter;
         padding-bottom: 10px;
         padding-left: 24px;
+        font-size: 14px;
     }
 
     ul {
@@ -38,7 +41,7 @@ aside {
         font-weight: 300;
         line-height: 32px;
         color: $primarycolor;
-        padding-left: 24px;
+        padding-left: 42px;
         width: 100%;
         display: inline-block;
     }
@@ -47,7 +50,6 @@ aside {
         .navlinksAnchor {
             background: $hovercolor;
             color: $secondarycolor;
-            font-weight: 500;
         }
     }
 }
