+++ docs/src/components/Search/index.scss
@@ -4,7 +4,7 @@
     padding: 24px;
     border-bottom: 1px $lightgrey solid;
     position: fixed;
-    width: calc(100% - 300px);
+    width: calc(100% - 250px);
     background: $white;
 }
 
@@ -24,10 +24,12 @@
 
     input {
         width: 100%;
-        border: 1px solid #c0c6cf;
+        border: 1px solid $borderColor;
         border-radius: 6px;
         height: 32px;
-        padding-left: 45px;
+        padding-left: 36px;
+        font-size: 14px;
+        color: $darkgrey;
     }
 
     input:focus {
